+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 1.0.0-pre.3
 
 - **BREAKING CHANGE**: Change options to separate, platform-specific object
-- You can now pass a headers and other options to the IO websocket client
+- You can now pass headers and other options to the IO web socket client
 - Backoff full jitter strategy for reconnection attempts
 - Add more metrics
 - Public Fake WebSocket client for testing